Graph-gen architectural changes to better support multiple telemetry vendors - simplify graph handler and push down telemetry client creation into vendor support - update API and push down telemetry client creation into vendor support - migrate shareable code from istio impl to telemetry:common.go - move handlers:graph_test.go to more appropriate graph/api:api_test.go package
https://github.com/kiali/kiali/issues/1679
